refactor: merge all common files in core/common folder
