+++ src/__mocks__/utils/image.mock.tsx
@@ -1,4 +1,4 @@
-jest.mock("../../common/utils/image", () => ({
+jest.mock("../../core/common/utils/image", () => ({
     bitmapToBlob: (bitmap: ImageBitmap): Blob => {
         return new Blob(['fake blob'], { type: 'image/png' });
     }
