--- src/common/utils/expressive.tsx
@@ -1,1 +1,0 @@
-export type Optional<T> = T | undefined;
