+++ src/core/assets/image/adapters/FileSystemImageRepository.test.tsx
@@ -1,5 +1,5 @@
 import { ImageAsset } from "@sparkengine";
-import { bitmapToBlob, WeakRef } from "../../../../common";
+import { bitmapToBlob, WeakRef } from "../../../../core/common";
 import { FileSystemImageRepository } from "./FileSystemImageRepository";
 import { createDirectoryHandleMock, FileSystemWritableFileStreamMock, setMockedFile } from "../../../../__mocks__/fs-api.mock";
 import { FakeBitmap } from "../../../../__mocks__/bitmap.mock";
