+++ src/core/common/dependencyInjection.tsx
@@ -28,5 +28,3 @@ export const DI = {
         return new config.di[typeName].concrete(...(deps || [])) as T;
     }
 };
-
-export type Factory<T extends abstract new (...args: any) => any> = (...args: ConstructorParameters<T>) => InstanceType<T>;
