+++ src/core/common/index.tsx
@@ -1,3 +1,4 @@
 export * from './FileSystemRepository';
 export * from './uuidToRgb';
 export * from './dependencyInjection';
+export * from './utils';
