+++ src/core/common/utils/function.tsx
