+++ src/core/common/utils/image.tsx
