+++ src/core/common/utils/weak-ref.tsx
