+++ src/core/project/adapters/FileSystemProjectRepository.tsx
@@ -1,4 +1,4 @@
-import { WeakRef } from "../../../common";
+import { WeakRef } from "../../../core/common";
 import { Project } from "../models";
 import { ProjectRepository } from "../ports";
