+++ src/core/project/models/Project.test.tsx
@@ -1,7 +1,7 @@
 import { Scene } from "@sparkengine";
 import { FileSystemSceneRepository } from "../../scene"
 import { Project } from "./Project"
-import { WeakRef } from "../../../common";
+import { WeakRef } from "../../../core/common";
 import { createDirectoryHandleMock, setMockedFile } from "../../../__mocks__/fs-api.mock";
 import testSceneJson from '../../../__mocks__/assets/test-scene.json';
